Clean up PHP 8.1 implementation
