+++ src/test/php/PDepend/Source/Language/PHP/Features/PHP80/PHPParserVersion80TestCase.php
@@ -60,7 +60,7 @@ abstract class PHPParserVersion80TestCase extends AbstractTestCase
     protected function createPHPParser(Tokenizer $tokenizer, Builder $builder, CacheDriver $cache)
     {
         return $this->getAbstractClassMock(
-            'PDepend\\Source\\Language\\PHP\\PHPParserVersion81',
+            'PDepend\\Source\\Language\\PHP\\AbstractPHPParser',
             [$tokenizer, $builder, $cache]
         );
     }
