+++ src/test/php/PDepend/Source/Language/PHP/Features/PHP81/ArrayUnpackingWithStringKeysTest.php
@@ -46,10 +46,9 @@ use PDepend\Source\AST\ASTReturnStatement;
 use PDepend\Source\AST\ASTTypeArray;
 
 /**
- * @covers \PDepend\Source\Language\PHP\PHPParserVersion81
+ * @covers \PDepend\Source\Language\PHP\AbstractPHPParser
  * @copyright 2008-2017 Manuel Pichler. All rights reserved.
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- *
  * @group unittest
  * @group php8.1
  */
