+++ src/test/php/PDepend/Source/Language/PHP/Features/PHP81/EnumTest.php
@@ -49,10 +49,9 @@ use PDepend\Source\AST\ASTMethod;
 use PDepend\Source\AST\ASTParameter;
 
 /**
- * @covers \PDepend\Source\Language\PHP\PHPParserVersion81
+ * @covers \PDepend\Source\Language\PHP\AbstractPHPParser
  * @copyright 2008-2017 Manuel Pichler. All rights reserved.
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- *
  * @group unittest
  * @group php8.1
  */
