+++ src/test/php/PDepend/Source/Language/PHP/Features/PHP81/ExplicitOctalNotationTest.php
@@ -44,7 +44,6 @@ use PDepend\Source\AST\ASTConstantDeclarator;
 use PDepend\Source\AST\ASTLiteral;
 
 /**
- * @covers \PDepend\Source\Language\PHP\PHPParserVersion81
  * @covers \PDepend\Source\Language\PHP\AbstractPHPParser
  * @copyright 2008-2017 Manuel Pichler. All rights reserved.
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
