+++ src/test/php/PDepend/Source/Language/PHP/Features/PHP82/TrueTypeTest.php
@@ -48,10 +48,9 @@ use PDepend\Source\AST\ASTScalarType;
 use PDepend\Source\AST\ASTVariableDeclarator;
 
 /**
- * @covers \PDepend\Source\Language\PHP\PHPParserVersion81
+ * @covers \PDepend\Source\Language\PHP\AbstractPHPParser
  * @copyright 2008-2017 Manuel Pichler. All rights reserved.
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- *
  * @group unittest
  * @group php8.2
  */
