+++ src/test/php/PDepend/Source/Language/PHP/PHPParserVersion73Test.php
@@ -184,7 +184,7 @@ class PHPParserVersion73Test extends AbstractTestCase
     protected function createPHPParser(Tokenizer $tokenizer, Builder $builder, CacheDriver $cache)
     {
         return $this->getAbstractClassMock(
-            'PDepend\\Source\\Language\\PHP\\PHPParserVersion81',
+            'PDepend\\Source\\Language\\PHP\\AbstractPHPParser',
             [$tokenizer, $builder, $cache]
         );
     }
